Image to Webp Converting
effort is 120 minutes

Almost all the online resources made examples with creating a model and adding image field to it and then converting that image which saved on database.

I think doing this conversion by using database is expensive and worthless. The requirement is only saying that this conversion must be done in backend. There is no need to store any information about the conversion.

Doing this converting feature with this method took more time than it must be but outcome of this effort is much more profitable than other solutions.

Note: In the next commits I will improve UI but now it works.