Updated Readme and Admin Panel View
effort is 40 minutes

In django auto_now fields are not displayed on admin panel because they are read only so i have added them manually and resolved a small bug :smile:

Hope you like this project. Even if we wont work together I really enjoyed and learn a lot with this project.